Riverbend: Add build information, and application information
We store the git hash inside of Build.h, and a human-rememberable
"codename". The codename can be whatever it wants to be, just some
preferably English word that isn't random characters and numbers.
This information can also be copied to the clipboard in the About menu.

The license is now also shown inside the About menu.