Riverbend: Correct scaling math, make scroll wheel less sensitive, add scale to view  menu.
Now we render the texture by the scale without the offset, and only by
the max of the texture (aka the bottom-right corner). This makes
everything feel much nicer, and also fixes a bug with texture that
weren't width == height. (which I never noticed because all my tests
were with that size)

Also adds the viewport scale to the view menu, to allow it to be
modified easier in some cases.

And a new Re-center button in the View menu, so you don't ever lose the
viewport!