LibXNA/Riverbend: Add support for compressed XNBs(!), restrict texture viewport loading
Compressed data segments are now read, parsed, and stored, uncompressed,
alongside the XNB object. This allows a viewport to use it later on when
displaying the primary asset :^)

Now that a lot more files can be loaded (aka, all compressed and
uncompressed ones), we now crash when loading some of these. Lets
avoid that by just not trying to render the ones we can't.